+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -25,7 +25,7 @@ object Libs {
         const val moshi = "1.15.1"
         const val okhttp = "4.12.0"
         const val retrofit = "2.11.0"
-        const val threetenbp = "1.6.9"
+        const val threetenbp = "1.7.0"
         const val truth = "1.4.4"
     }