+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -6,8 +6,8 @@ object GradlePlugins {
 
     private object Versions {
         const val dokka = "1.9.20"
-        const val kotlin = "1.9.24"
-        const val ksp = "1.9.24-1.0.20"
+        const val kotlin = "2.0.21"
+        const val ksp = "2.0.21-1.0.27"
         const val versions = "0.51.0"
     }